Fix warnings  * a lot of functions are still broken and need retesting

+++ kernel/src/math/os_long_int_negate.c
@@ -7,7 +7,7 @@ void _os_long_int_negate(int* ax, int* bx, int* cx, int* dx, int* si, int* di)
 
 void os_long_int_negate(int* hi, int* lo)
 {
-    long temp = (*lo) | ((*hi) << 16);
+    uint32_t temp = (uint32_t) (*lo) | ((uint32_t) (*hi) << 16);
     temp = -temp;
     *lo = temp & 0x0000FFFF;
     *hi = (temp >> 16) & 0x0000FFFF;
